requirements.txt
tornado==6.0.3
awscli
plumbum
docker
ipython
sentry-sdk
real-easypy==0.4.3
requests
pyfiglet
